Controlling the cs lines separately

+++ LSM9DS1_Types.h
@@ -201,8 +201,10 @@ struct gyroSettings
 struct deviceSettings
 {
     uint8_t commInterface; // Can be I2C, SPI 4-wire or SPI 3-wire
-    uint8_t agAddress;	// I2C address or SPI CS pin
-	uint8_t mAddress;	// I2C address or SPI CS pin
+    //uint8_t agAddress;	// I2C address or SPI CS pin
+	//uint8_t mAddress;	// I2C address or SPI CS pin
+    int agAddress;	// I2C address or SPI CS pin
+    int mAddress;	// I2C address or SPI CS pin
 };
 
 struct accelSettings
